Fix LVV-T1802 to run from start to finish on TTS.
- Fix typos to make the notebook run
    - remove extra .get() s etc
    - fix incorrect assignment between variable names in velocity commands
    - Fix some incorrect ‘accepted’ vs ‘rejected’ text.

- I added an explanatory sentence:’In CompenstationMode a LUT is used to
compensate during the movements.’

- Correct the procedure to read and printout the summary state
variables.

- Fix the cmd_configureLimits cells that were missing ‘await’ commands,
Now the command is really issued.

- Before the acceleration and velocity commands, add a new cell to reset
the configuration limits to their defaults as seen on the EUI.

- Fix the hexapod_csc.cmd_configureAcceleration.set_start to use an
‘acceleration’ keyword.

- Remove the final stress test loop and rename it to be Hexapod Envelope
test

Note: There are still notes from Bruno in the notebook about new
things that need to be implemented.

Note: There is some confusing text from Holger that came from the Jira
test case:

‘The following steps define what the Jupyter Notebook for this test
case implements. Executing the Jupyter notebook is the only actual
command and control step that needs to be executed. Transition the
state machine into disabledState to publish telemetry.’

To check for the future: The association between calibrated and raw
etc strut positions do not seem to agree well with the printed out
hexapod positions.