Add example notebook for retrieving block information.
This notebook is for SITCOM-1209.  It is an example of how
you can use the tma summit utils to retrieve information on
what blocks were run on a certain range of dayObs.  It also
produces a more complete example report. It makes use of
awkward arrays and it's panda interface to show a pandas
table with variable length lists of block numbers in a
column.